+++ js/loader.js
@@ -1,6 +1,6 @@
 /*global loadFiles dashtype */
 
-var _DASHTICZ_VERSION = 130;
+var _DASHTICZ_VERSION = 131;
 var head = document.getElementsByTagName('head')[0],
     script = document.createElement('script');